+++ src/main/java/tamanna/challange/interviews/repository/AvailableInterviewDatesRepository.java
@@ -8,7 +8,7 @@ import tamanna.challange.interviews.model.Schedule.AvailableInterviewDates;
 import java.util.List;
 
 @Repository
-public interface AvailableInterviewDatesRepository extends JpaRepository<AvailableInterviewDates,Integer> {
+public interface AvailableInterviewDatesRepository extends JpaRepository<AvailableInterviewDates,Long> {
             @Query(value = "SELECT * FROM AVAILABLE_INTERVIEW_DATES WHERE INTERVIEWER_ID = :interviewerId ", nativeQuery=true )
-            List< AvailableInterviewDates > getAvailableDatesByInterviewer(Integer interviewerId);
+            List< AvailableInterviewDates > getAvailableDatesByInterviewer(Long interviewerId);
 }
