+++ src/main/java/tamanna/challange/interviews/repository/RequestedInterviewDatesRepository.java
@@ -8,8 +8,8 @@ import tamanna.challange.interviews.model.Schedule.RequestedInterviewDates;
 import java.util.List;
 
 @Repository
-public interface RequestedInterviewDatesRepository extends JpaRepository<RequestedInterviewDates,Integer> {
+public interface RequestedInterviewDatesRepository extends JpaRepository<RequestedInterviewDates,Long> {
     @Query(value = "SELECT * FROM REQUESTED_INTERVIEW_DATES WHERE CANDIDATE_ID = :candidateId ", nativeQuery=true )
-    List<RequestedInterviewDates> getRequestedDatesByCandidate(Integer candidateId);
+    List<RequestedInterviewDates> getRequestedDatesByCandidate(Long candidateId);
 
 }
